+++ python/src/lib.rs
@@ -134,6 +134,10 @@ fn frametransform(_py: Python, m: &Bound<'_, PyModule>) -> PyResult<()> {
         .unwrap();
     m.add_function(wrap_pyfunction!(pyft::gcrf_to_itrf_state, m)?)
         .unwrap();
+    m.add_function(wrap_pyfunction!(pyft::itrf_to_gcrf_state_approx, m)?)
+        .unwrap();
+    m.add_function(wrap_pyfunction!(pyft::gcrf_to_itrf_state_approx, m)?)
+        .unwrap();
     m.add_function(wrap_pyfunction!(pyft::qmod2gcrf, m)?)
         .unwrap();
     m.add_function(wrap_pyfunction!(pyft::qtod2mod_approx, m)?)